refactor(app): scope identity spec document crypto to its own module
Move the identity-spec setup document crypto out of identity.rs into
identity/spec_document.rs so its transitional dead-code allowance covers
only the staged crypto rather than the unrelated principal helpers that
share the module. This mirrors credentials/encryption.rs, which stages
its own primitives behind a file-scoped allowance the same way.

Share the static and rotating CredentialKeyProvider test fixtures from
credentials::encryption::test_support instead of keeping a second copy
in the identity tests, so rotation and error semantics stay consistent
as either suite evolves.
